refactor: component optimization
