+++ .vscode/settings.json
@@ -25,11 +25,14 @@
     "backtop",
     "cascader",
     "codesandbox",
+    "maxcharacter",
+    "maxlength",
     "miniprogram",
     "nuxt",
     "popconfirm",
     "popperjs",
     "precaching",
+    "sider",
     "tdesign",
     "tdoc",
     "tmenu",
