+++ packages/components/empty/components/SuccessSvg.tsx
@@ -2,8 +2,8 @@ import { defineComponent } from 'vue';
 
 export default defineComponent({
   name: 'SuccessSvg',
-  render() {
-    return (
+  setup() {
+    return () => (
       <svg width="1em" height="1em" viewBox="0 0 48 48" fill="none" xmlns="http://www.w3.org/2000/svg">
         <path
           fill-rule="evenodd"
