+++ packages/components/form/utils/form-model.ts
@@ -15,7 +15,7 @@ import {
   AllValidateResult,
   ValidateResultType,
   CustomValidateResolveType,
-} from './type';
+} from '../type';
 
 // `{} / [] / '' / undefined / null` 等内容被认为是空； 0 和 false 被认为是正常数据，部分数据的值就是 0 或者 false
 export function isValueEmpty(val: ValueType): boolean {
