+++ packages/components/grid/row.tsx
@@ -1,14 +1,13 @@
 import { defineComponent, provide, computed, toRefs, reactive } from 'vue';
 import props from './row-props';
-import { useRowSize, getRowClasses, RowProviderType, calcRowStyle } from './common';
+import { useRowSize } from './hooks';
+import { getRowClasses, RowProviderType, calcRowStyle } from './utils';
 import { usePrefixClass } from '../hooks/useConfig';
 import { useTNodeJSX } from '../hooks/tnode';
 
 export default defineComponent({
   name: 'TRow',
-
   props,
-
   setup(props) {
     const { gutter } = toRefs(props);
     const renderTNodeJSX = useTNodeJSX();
