+++ packages/components/guide/utils/index.ts
@@ -1,6 +1,4 @@
-import { scrollToParentVisibleArea } from './getScrollParent';
-import getRelativePosition from './getRelativePosition';
-import getTargetElm from './getTargetElm';
-import scrollToElm from './scrollToElm';
-
-export { scrollToParentVisibleArea, getRelativePosition, getTargetElm, scrollToElm };
+export * from './getScrollParent';
+export * from './getRelativePosition';
+export * from './getTargetElm';
+export * from './scrollToElm';
