+++ packages/components/image-viewer/consts/index.ts
