+++ packages/components/image-viewer/hooks/index.ts
@@ -1,6 +1,6 @@
 import { positiveSubtract, positiveAdd } from '@tdesign/common-js/input-number/number';
 import { ref, watch } from 'vue';
-import { ImageScale } from './type';
+import { ImageScale } from '../type';
 import { throttle } from 'lodash-es';
 
 interface InitTransform {
