+++ packages/components/image-viewer/image-viewer.tsx
@@ -12,7 +12,7 @@ import TImageItem from './base/ImageItem';
 import TImageViewerIcon from './base/ImageModalIcon';
 import TImageViewerModal from './base/ImageViewerModal';
 import TImageViewerUtils from './base/ImageViewerUtils';
-import { EVENT_CODE } from './const';
+import { EVENT_CODE } from './consts';
 import { useMirror, useRotate, useScale } from './hooks';
 import props from './props';
 import { TdImageViewerProps } from './type';
