+++ packages/components/input/input.tsx
@@ -10,9 +10,7 @@ import { useReadonly } from '../hooks/useReadonly';
 import { useConfig, usePrefixClass, useCommonClassName } from '../hooks/useConfig';
 import { useGlobalIcon } from '../hooks/useGlobalIcon';
 import { useTNodeJSX } from '../hooks/tnode';
-import useInput from './useInput';
-import useInputEventHandler from './useInputEventHandler';
-import useInputWidth from './useInputWidth';
+import { useInput, useInputEventHandler, useInputWidth } from './hooks';
 import { getValidAttrs } from '@tdesign/common-js/utils/helper';
 
 export default defineComponent({
@@ -34,7 +32,6 @@ export default defineComponent({
       default: false,
     },
   },
-
   setup(props, { expose }) {
     const { globalConfig } = useConfig('input');
     const { BrowseIcon, BrowseOffIcon, CloseCircleFilledIcon } = useGlobalIcon({
