+++ packages/components/table/editable-cell.tsx
@@ -15,7 +15,7 @@ import {
 import { TableClassName } from './hooks/useClassName';
 import { useGlobalIcon } from '../hooks/useGlobalIcon';
 import { renderCell } from './tr';
-import { validate } from '../form/form-model';
+import { validate } from '../form/utils/form-model';
 import log from '@tdesign/common-js/log/index';
 import { AllValidateResult } from '../form/type';
 import { on, off } from '../utils/dom';
