+++ packages/components/textarea/textarea.tsx
@@ -13,7 +13,7 @@ import {
 import { isObject, merge, omit } from 'lodash-es';
 
 import { calcTextareaHeight } from './utils';
-import { FormItemInjectionKey } from '../form/const';
+import { FormItemInjectionKey } from '../form/consts';
 import setStyle from '@tdesign/common-js/utils/setStyle';
 import { getCharacterLength } from '@tdesign/common-js/utils/helper';
 
@@ -23,7 +23,7 @@ import { useDisabled } from '../hooks/useDisabled';
 import { useReadonly } from '../hooks/useReadonly';
 import { useTNodeJSX } from '../hooks/tnode';
 import { usePrefixClass, useCommonClassName } from '../hooks/useConfig';
-import useLengthLimit from '../input/useLengthLimit';
+import { useLengthLimit } from '../input/hooks/useLengthLimit';
 
 import props from './props';
 import type { TextareaValue, TdTextareaProps } from './type';
